// src/api/generated/model/expense.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Go Funny API
 * OpenAPI spec version: 0.1.0
 */
import type { ExpenseSplit } from './expenseSplit';
import type { SplitType } from './splitType';

export interface Expense {
  id: string;
  tripId: string;
  title: string;
  amount: number;
  date: string;
  splitType: SplitType;
  /** @nullable */
  payerMembershipId?: string | null;
  /** @nullable */
  note?: string | null;
  createdAt: string;
  splits?: ExpenseSplit[];
}